refactor: separate waitlist types and add status field to support invites
